# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Silk Way API",
        "description": "REST API для мобильного приложения Silk Way. Авторизация через Bearer токен (Sanctum).",
        "contact": {
            "email": "admin@silkway.kz"
        },
        "version": "1.9.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Список всех активных машин",
                "operationId": "5a4232faea14f4b6d28da69ed12170a1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список машин",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Car"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "429": {
                        "description": "Слишком много запросов (120/мин)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Добавить машину",
                "description": "Только для водителей.",
                "operationId": "628a4198ebf8866101714ec6c8feea4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "brand",
                                    "model",
                                    "license_plate",
                                    "max_weight",
                                    "trailer_length",
                                    "trailer_width",
                                    "trailer_height",
                                    "trailer_type"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string",
                                        "example": "Volvo"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "FH16"
                                    },
                                    "license_plate": {
                                        "type": "string",
                                        "example": "A123BC"
                                    },
                                    "max_weight": {
                                        "type": "number",
                                        "example": 20
                                    },
                                    "trailer_length": {
                                        "type": "number",
                                        "example": 13.6
                                    },
                                    "trailer_width": {
                                        "type": "number",
                                        "example": 2.4
                                    },
                                    "trailer_height": {
                                        "type": "number",
                                        "example": 2.7
                                    },
                                    "trailer_type": {
                                        "type": "string",
                                        "enum": [
                                            "tral",
                                            "refrigerator",
                                            "tent"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Машина добавлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ только для водителей"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "429": {
                        "description": "Слишком много запросов (120/мин)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cars/my": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Машины текущего водителя",
                "operationId": "b9abe373e3080aa51fac9057ea92c141",
                "responses": {
                    "200": {
                        "description": "Список машин водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Car"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Детали машины",
                "operationId": "018ed5bf0dd975a8e657ae8e0c63a7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали машины",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Обновить машину (form-data с _method=PUT)",
                "operationId": "531ad7a5d36382ed878ec85c7a56bd5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "brand",
                                    "model",
                                    "license_plate",
                                    "max_weight",
                                    "trailer_length",
                                    "trailer_width",
                                    "trailer_height",
                                    "trailer_type"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "license_plate": {
                                        "type": "string"
                                    },
                                    "max_weight": {
                                        "type": "number"
                                    },
                                    "trailer_length": {
                                        "type": "number"
                                    },
                                    "trailer_width": {
                                        "type": "number"
                                    },
                                    "trailer_height": {
                                        "type": "number"
                                    },
                                    "trailer_type": {
                                        "type": "string",
                                        "enum": [
                                            "tral",
                                            "refrigerator",
                                            "tent"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Машина обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Нет доступа"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "429": {
                        "description": "Слишком много запросов (120/мин)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Удалить машину",
                "operationId": "b03c354ae2b0640b709a68317865f8e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Машина удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Машина удалена."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Нет доступа"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "429": {
                        "description": "Слишком много запросов (120/мин)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cars/{id}/toggle": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Активировать / деактивировать машину",
                "operationId": "9b73773e2710c38922bec0e453811877",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус изменён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нет доступа"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Car": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand": {
                        "type": "string",
                        "example": "Volvo"
                    },
                    "model": {
                        "type": "string",
                        "example": "FH16"
                    },
                    "license_plate": {
                        "type": "string",
                        "example": "A123BC"
                    },
                    "max_weight": {
                        "type": "number",
                        "example": 20
                    },
                    "trailer_type": {
                        "type": "string",
                        "enum": [
                            "tral",
                            "refrigerator",
                            "tent"
                        ]
                    },
                    "trailer_length": {
                        "type": "number",
                        "example": 13.6
                    },
                    "trailer_width": {
                        "type": "number",
                        "example": 2.4
                    },
                    "trailer_height": {
                        "type": "number",
                        "example": 2.7
                    },
                    "trailer_volume": {
                        "type": "number",
                        "example": 88.13
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите токен (без слова Bearer — оно добавляется автоматически)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}